requirements.txt
alembic==0.9.4
amqp==5.0.6
billiard==3.6.4.0
celery==5.1.2
certifi==2017.7.27.1
cffi==1.14.6
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
cryptography==3.4.8
Flask==0.10.1
Flask-Migrate==2.7.0
Flask-Script==2.0.5
Flask-Session==0.3.1
Flask-SQLAlchemy==2.2
Flask-WTF==0.14.2
idna==2.5
itsdangerous==0.24
Jinja2==2.9.6
kombu==5.1.0
Mako==1.0.7
MarkupSafe==2.0.1
olefile==0.44
Pillow==8.3.1
prompt-toolkit==3.0.19
pycparser==2.20
pycryptodomex==3.9.4
PyMySQL==1.0.2
pyOpenSSL==19.1.0
python-alipay-sdk==3.0.4
python-dateutil==2.6.1
python-editor==1.0.3
pytz==2021.1
qiniu==7.1.4
redis==2.10.5
requests==2.18.3
ronglian-sms-sdk==1.0.0
six==1.10.0
SQLAlchemy==1.1.12
urllib3==1.22
uWSGI==2.0.19.1
vine==5.0.0
wcwidth==0.2.5
Werkzeug==0.12.2
WTForms==2.1
